Move components without changing Transifex paths
