+++ test/components/dataset/entities.spec.js
@@ -1,5 +1,5 @@
 import Selectable from '../../../src/components/selectable.vue';
-import OdataAnalyze from '../../../src/components/submission/analyze.vue';
+import OdataAnalyze from '../../../src/components/odata/analyze.vue';
 
 import testData from '../../data';
 import { load } from '../../util/http';
